# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalCreateRequest": {
        "description": "The restore plan a second administrator is asked to review.",
        "properties": {
          "restore_operation_id": {
            "$ref": "#/components/schemas/PydanticObjectId"
          }
        },
        "required": [
          "restore_operation_id"
        ],
        "title": "ApprovalCreateRequest",
        "type": "object"
      },
      "ApprovalDecisionRequest": {
        "description": "An optional reason recorded with an approval decision.",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "ApprovalDecisionRequest",
        "type": "object"
      },
      "ApprovalListResponse": {
        "description": "Paginated approval requests.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApprovalResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "ApprovalListResponse",
        "type": "object"
      },
      "ApprovalResponse": {
        "description": "An approval request as the restore page renders it.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By Email"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "delete_count": {
            "title": "Delete Count",
            "type": "integer"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "object_count": {
            "title": "Object Count",
            "type": "integer"
          },
          "plan_hash": {
            "title": "Plan Hash",
            "type": "string"
          },
          "requested_by_email": {
            "title": "Requested By Email",
            "type": "string"
          },
          "restore_operation_id": {
            "title": "Restore Operation Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApprovalStatus"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "triggered_rules": {
            "items": {
              "$ref": "#/components/schemas/TriggeredRuleResponse"
            },
            "title": "Triggered Rules",
            "type": "array"
          }
        },
        "required": [
          "id",
          "restore_operation_id",
          "status",
          "triggered_rules",
          "requested_by_email",
          "decided_by_email",
          "decided_at",
          "decision_reason",
          "expires_at",
          "plan_hash",
          "summary",
          "object_count",
          "delete_count",
          "created_at"
        ],
        "title": "ApprovalResponse",
        "type": "object"
      },
      "ApprovalRule": {
        "description": "Policy rule that made a plan require approval.",
        "enum": [
          "organization_scope",
          "exact_mode_deletes",
          "object_count_threshold",
          "sensitive_object_type"
        ],
        "title": "ApprovalRule",
        "type": "string"
      },
      "ApprovalStatus": {
        "description": "Approval request lifecycle.",
        "enum": [
          "pending",
          "approved",
          "rejected",
          "expired",
          "invalidated"
        ],
        "title": "ApprovalStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PydanticObjectId": {
        "example": "5eb7cf5a86d9755df3a6c593",
        "maxLength": 24,
        "minLength": 24,
        "pattern": "^[0-9a-f]{24}$",
        "type": "string"
      },
...
      "TriggeredRuleResponse": {
        "description": "One policy rule with the evidence that triggered it.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "rule": {
            "$ref": "#/components/schemas/ApprovalRule"
          }
        },
        "required": [
          "rule",
          "detail"
        ],
        "title": "TriggeredRuleResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CsrfToken": {
        "description": "Double-submit CSRF token. Required on every unsafe method of a cookie session.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      },
      "SessionCookie": {
        "description": "Browser session issued by `POST /auth/login`. Unsafe methods must also send the `cg_csrf` cookie value in the `X-CSRF-Token` header.",
        "in": "cookie",
        "name": "cg_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "summary": "Configuration history, point-in-time recovery, and impact monitoring for Juniper Mist.",
    "title": "Mist Config Guardian",
    "version": "0.2.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/organizations/{organization_id}/approvals": {
      "get": {
        "description": "List approval requests newest first, optionally filtered by status.",
        "operationId": "list_approvals_api_v1_organizations__organization_id__approvals_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApprovalStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "List Approvals",
        "tags": [
          "Approvals"
        ]
      },
      "post": {
        "description": "Ask a second administrator to review one restore plan.",
        "operationId": "request_approval_api_v1_organizations__organization_id__approvals_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Request Approval",
        "tags": [
          "Approvals"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/approvals/{approval_id}": {
      "get": {
        "description": "Return one approval request with a freshly reconciled status.",
        "operationId": "get_approval_api_v1_organizations__organization_id__approvals__approval_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Get Approval",
        "tags": [
          "Approvals"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/approvals/{approval_id}/approve": {
      "post": {
        "description": "Approve one restore plan as a second administrator.",
        "operationId": "approve_restore_api_v1_organizations__organization_id__approvals__approval_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Approve Restore",
        "tags": [
          "Approvals"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/approvals/{approval_id}/reject": {
      "post": {
        "description": "Reject one restore plan as a second administrator.",
        "operationId": "reject_restore_api_v1_organizations__organization_id__approvals__approval_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Reject Restore",
        "tags": [
          "Approvals"
        ]
      }
    },
...
  },
...
}